pyFAST_new/requirements.txt
numpy
pandas
matplotlib
tqdm
pytorch>=2.0
transformers
datasets
sympy
scipy
